Move filename dialog creation to handler
